Finished 1.0 version

+++ auric.sh
@@ -23,30 +23,30 @@ VERSION="1.0.0"
 # Local AUR git repo
 AURDIR="$HOME/.AUR"
 
+# AUR package info URL
+AUR_INFO_URL="https://aur.archlinux.org/rpc/?v=5&type=info&arg[]="
+
 # AUR package search URL
-AURURL="https://aur.archlinux.org/rpc/?v=5&type=info&arg[]="
+AUR_SRCH_URL="https://aur.archlinux.org/rpc/?v=5&type=search&by=name&arg="
 
 # GIT URL for AUR repos. %s will be replaced with package name
-GITURL="https://aur.archlinux.org/%s.git"
+GIT_AUR_URL="https://aur.archlinux.org/%s.git"
 
-# Successful git pull result - no spaces or punctuation
-GITRES="alreadyuptodate"
-
-# Whether to resolve pacman dependencies during download.
-# Makepkg does this so its unlikely this will need to be true
-USEPAC=false
+# Successful git pull result - lowercase with no spaces, dashes, or punctuation
+GIT_RES_STR="alreadyuptodate"
 
 # Will contain the list of all installed packages. 
 # This gets set automatically
-LOCALPKGS=""
+LOCAL_PKGS=""
 
-# Name of installed JSON parser. This gets set automatically.
-JSONPARSER=""
+# Name of installed JSON parser.
+# This gets set automatically.
+JSON_PARSER=""
 
 # Array containg all successfully downloaded packages.
 # If this contains package names AURIC will prompt
 # user to install after downloading
-TOINSTALL=()
+TO_INSTALL=()
 
 # ----------------------------------------------------------------------------------
 
@@ -131,8 +131,8 @@ download() {
     validate_pkgname "$1"
 
     # Get a list of all installed packages
-    if [[ "${LOCALPKGS}" == "" ]]; then
-        LOCALPKGS=$(pacman -Sl)
+    if [[ "${LOCAL_PKGS}" == "" ]]; then
+        LOCAL_PKGS=$(pacman -Sl)
     fi
 
     # Move into the AUR folder
@@ -141,10 +141,10 @@ download() {
     for PKG in $@; do
 
         # Fetch the JSON data associated with the submitted package name
-        curl_result=$(curl -fsSk "${AURURL}${PKG}")
+        curl_result=$(curl -fsSk "${AUR_INFO_URL}${PKG}")
 
         # Parse the result using the installed JSON parser
-        if [[ $JSONPARSER == 'jq' ]]; then
+        if [[ $JSON_PARSER == 'jq' ]]; then
             json_result=$(echo "$curl_result" | jq -r '.results')
         else
             json_result=$(echo "$curl_result" | jshon -e results)
@@ -153,7 +153,7 @@ download() {
         # Did the package query return a valid result?
         # If not it means the package is not in the AUR
         if [[ $json_result == "[]" ]]; then
-            echo -e "${red}MISSING:${reset} ${PKG} not in AUR. Makepkg will install it via pacman if needed"
+            echo -e "${red}MISSING:${reset} ${cyan}${PKG}${reset} ${red}not in AUR.${reset} ${yellow}Makepkg will resolve all non-AUR dependencies${reset}"
         else
 
             # If a folder with the package name exists in the local repo we skip it
@@ -165,7 +165,7 @@ download() {
             echo -e "${yellow}CLONING:${reset} $PKG"
 
             # Assemble the git package URL
-            printf -v URL "$GITURL" "$PKG"
+            printf -v URL "$GIT_AUR_URL" "$PKG"
 
             # Clone it
             git clone $URL 2> /dev/null
@@ -185,10 +185,10 @@ download() {
             fi
 
             # Add the package to the install array
-            TOINSTALL+=("$PKG")
+            TO_INSTALL+=("$PKG")
 
             # Get the package dependencies using installed json parser
-            if [[ $JSONPARSER == 'jq' ]]; then
+            if [[ $JSON_PARSER == 'jq' ]]; then
                 has_depends=$(echo "$curl_result" | jq -r '.results[0].Depends') 
             else
                 has_depends=$(echo "$curl_result" | jshon -e results -e 0 -e Depends)
@@ -197,7 +197,7 @@ download() {
             # If there is a result, recurisvely call this function with the dependencies
             if [[ $has_depends != "[]" ]] && [[ $has_depends != null ]]; then
 
-                if [[ $JSONPARSER == 'jq' ]]; then
+                if [[ $JSON_PARSER == 'jq' ]]; then
                     dependencies=$(echo "$curl_result" | jq -r '.results[0].Depends[]') 
                 else
                     dependencies=$(echo "$curl_result" | jshon -e results -e 0 -e Depends -a -u)
@@ -210,10 +210,9 @@ download() {
                     # Some dependencies have minimum version requirements
                     # which screws up the package name
                     depend=$(echo $depend | sed "s/[>=].*//")
-                    depend=${depend// /} # remove spaces just in case
 
                     # See if the dependency is already installed
-                    echo "$LOCALPKGS" | grep "$depend" > /dev/null
+                    echo "$LOCAL_PKGS" | grep "$depend" > /dev/null
 
                     # Download it
                     if [[ "$?" == 1 ]]; then
@@ -236,11 +235,7 @@ install() {
     PKG=$1
 
     if [[ ! -d "${AURDIR}/${PKG}" ]]; then
-        echo -e "${red}Error: Package does not exist${reset}"
-        echo
-        echo "See help page:"
-        echo
-        echo -e "   ${yel}\$${reset}   ${grn}auric --help${reset}"
+        echo -e "${red}MISSING:${reset} Package does not exist in local repo"
         echo
         exit 1
     fi
@@ -254,7 +249,7 @@ install() {
         exit 1
     fi
 
-    echo -e "Running makepkg on ${cyan}${PKG}${reset}"
+    echo -e "RUNNING MAKEPKG ON ${cyan}${PKG}${reset}"
     echo
 
     # MAKEPKG FLAGS
@@ -280,16 +275,18 @@ update() {
             echo -e "${yellow}PULLING:${reset} ${DIR:2}"
 
             RESULT=$(git pull 2> /dev/null)
+
+            # Format the result string for reliability
             RESULT=${RESULT//./} # remove periods
             RESULT=${RESULT//-/} # remove dashes
             RESULT=${RESULT// /} # remove spaces
             RESULT=${RESULT,,}   # lowercase
 
-            if [[ $RESULT != "$GITRES" ]]; then
+            if [[ $RESULT != "$GIT_RES_STR" ]]; then
                 echo -e "${red}NEW VER:${reset} ${cyan}${DIR:2}${reset} has been updated and must be reinstalled"
 
                 # Add the package to the install array
-                TOINSTALL+=("${DIR:2}")                
+                TO_INSTALL+=("${DIR:2}")                
             else
                 echo -e "${green}CURRENT:${reset} ${DIR:2} is up to date"
             fi
@@ -308,16 +305,18 @@ update() {
         cd ${AURDIR}/${PKG} || exit
 
         RESULT=$(git pull 2> /dev/null)
+
+        # Format the result string for reliability
         RESULT=${RESULT//./} # remove periods
         RESULT=${RESULT//-/} # remove dashes
         RESULT=${RESULT// /} # remove spaces
         RESULT=${RESULT,,}   # lowercase
 
-        if [[ $RESULT != "$GITRES" ]]; then
+        if [[ $RESULT != "$GIT_RES_STR" ]]; then
             echo -e "${red}NEW VER:${reset} ${cyan}${PKG}${reset} has been updated and must be reinstalled"
 
             # Add the package to the install array
-            TOINSTALL+=("$PKG")    
+            TO_INSTALL+=("$PKG")    
         else
             echo -e "${green}CURRENT:${reset} ${PKG} is up to date"
         fi
@@ -327,8 +326,40 @@ update() {
 
 # ----------------------------------------------------------------------------------
 
+# AUR package name search
 search() {
-  curl -s "https://aur.archlinux.org/rpc/?v=5&type=search&arg=$1" | jshon -a -e Name -u
+
+    local json_result
+    local PKG
+    PKG=$1
+
+    # Fetch the JSON data associated with the package name search
+    curl_result=$(curl -fsSk "${AUR_SRCH_URL}${PKG}")
+
+    # Parse the result using the installed JSON parser
+    if [[ $JSON_PARSER == 'jq' ]]; then
+        json_result=$(echo "$curl_result" | jq -r '.results[] .Name')
+    else
+        json_result=$(echo "$curl_result" | jshon -e results -a -e Name -u)
+    fi
+
+    if [[ $json_result == "[]" ]] || [[ $json_result == null ]]; then
+        echo -e "${red}NO RESULTS:${reset} No results for the term \"${cyan}${PKG}${reset}\""
+    else
+
+        echo "SEARCH RESULTS"
+        echo
+        for res in $json_result; do
+
+            # Capture the search term and surround it with %s
+            # so we can use printf to replace with color variables
+            res=$(echo "$res" | sed "s/\(.*\)\(${PKG}\)\(.*\)/\1%s\2%s\3/")
+
+            printf -v res "$res" "${cyan}" "${reset}"
+
+            echo -e " $res"
+        done
+    fi
 }
 
 # ----------------------------------------------------------------------------------
@@ -343,7 +374,7 @@ migrate() {
         PKG=${PKG// /}
         download "$PKG"
     done
-    TOINSTALL=()
+    TO_INSTALL=()
 }
 
 # ----------------------------------------------------------------------------------
@@ -372,37 +403,52 @@ remove() {
         exit 1
     fi
 
-    sudo pacman -Rsc $PKG
-    rm -rf ${AURDIR}/${PKG}
-
+    echo -e "Are you sure you want to remove the following package?"
     echo
-    echo -e "${red}${PKG} has been removed${reset}"
+    echo -e " ${cyan}${PKG}${reset}"
+    echo
+    read -p "ENTER [Y/n] " CONSENT
+
+    if [[ $CONSENT =~ [y|Y] ]]; then
+
+        sudo pacman -Rsc $PKG
+        rm -rf ${AURDIR}/${PKG}
+
+        echo
+        echo -e "${red}REMOVED:${reset} ${PKG}"
+    else
+        echo
+        echo "Goodbye..."
+    fi
 }
 
 # ----------------------------------------------------------------------------------
 
 # Is jq or jshon installed? 
 if command -v jq &>/dev/null; then
-    JSONPARSER="jq"
+    JSON_PARSER="jq"
 elif command -v jshon &>/dev/null; then
-    JSONPARSER="jshon"
+    JSON_PARSER="jshon"
 else
-    echo -e "${red}Error: No JSON parser installed${reset}"
+    echo -e "${red}DEPENDENCY ERROR:${reset} No JSON parser installed"
     echo
-    echo "This script requires either jq or jshon."
+    echo "This script requires either jq or jshon"
     echo
     exit 1
 fi
 
 # Is curl installed?
 if ! command -v curl &>/dev/null; then
-    echo -e "${red}Error: This script requires Curl${reset}"
+    echo -e "${red}DEPENDENCY ERROR:${reset} Curl not installed"
+    echo
+    echo "This script requires Curl to retrieve search results and package info"
     echo
     exit 1
 fi
 
 # No arguments, we show help
 if [[ -z "$1" ]]; then
+    echo -e "${yellow}HELP MENU${reset}"
     help
 fi
 
@@ -445,13 +491,13 @@ case "$CMD" in
 esac
 
 
-# If the TOINSTALL array contains package names we offer to install them
-if [[ ${#TOINSTALL[@]} -gt 0 ]]; then
+# If the TO_INSTALL array contains package names we offer to install them
+if [[ ${#TO_INSTALL[@]} -gt 0 ]]; then
 
     echo
     echo "Would you like to install the packages you downloaded?" 
     echo
-    for PKG in ${TOINSTALL[@]}; do
+    for PKG in ${TO_INSTALL[@]}; do
         echo -e "  ${cyan}${PKG}${reset}"
     done
     echo
@@ -463,8 +509,8 @@ if [[ ${#TOINSTALL[@]} -gt 0 ]]; then
         # Run through the install array in reverse order.
         # This helps ensure that dependencies get installed first.
         # The order doesn't matter after updating, only downloading
-        for (( i=${#TOINSTALL[@]}-1 ; i>=0 ; i-- )) ; do
-            install "${TOINSTALL[i]}"
+        for (( i=${#TO_INSTALL[@]}-1 ; i>=0 ; i-- )) ; do
+            install "${TO_INSTALL[i]}"
         done
 
     else
